Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 93ms, Memory: 9MB, Status: Wrong Answer - CodeTree

+++ 241124/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -47,9 +47,9 @@ class GameSol {
 
     public void init() throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        size = Integer.parseInt(br.readLine());
+        size = Long.parseLong(br.readLine());
         StringTokenizer st = new StringTokenizer(br.readLine());
-        start = Integer.parseInt(st.nextToken());
-        end = Integer.parseInt(st.nextToken());
+        start = Long.parseLong(st.nextToken());
+        end = Long.parseLong(st.nextToken());
     }
 }
